+++ crates/matrix-sdk-base/Changelog.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - Rename `RoomType` to `RoomState`
+- Add `RoomInfo::state` accessor
 
 ## 0.5.1